Добавил ViewModel на MainScreen
